refactor: refactor passport data types
